+++ blog/.gitignore
@@ -16,6 +16,7 @@ build/
 bin/
 !**/src/main/**/bin/
 !**/src/test/**/bin/
+application.yml
 
 ### IntelliJ IDEA ###
 .idea
