--- blog/src/main/java/efub/session/blog/post/repository/PostRepository.java
@@ -1,10 +1,0 @@
-package efub.session.blog.post.repository;
-
-import efub.session.blog.post.domain.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-    Optional<Post> findByPostIdAndWriter_AccountId(Long postId, Long accountId);
-}
